HoangAnh_Update_code.
sửa về tính dame tương sinh tương khắc cho chuẩn và hoạt động lại UI máu Enemy.
