+++ Assets/Game/Scripts/UI/HealEnemy/FollowEnemyHealthBar.cs
@@ -2,14 +2,14 @@ using UnityEngine;
 
 public class FollowEnemyHealthBar : MonoBehaviour
 {
-    public Transform enemyTransform; 
-    public Vector3 offset = new Vector3(0, 1, 0); 
+    public GameObject enemyTransform;
+    public Vector3 offset = new Vector3(0, 1, 0);
 
     void Update()
     {
         if (enemyTransform != null)
         {
-            Vector3 enemyScreenPos = Camera.main.WorldToScreenPoint(enemyTransform.position);
+            Vector3 enemyScreenPos = Camera.main.WorldToScreenPoint(enemyTransform.transform.position);
 
             transform.position = enemyScreenPos + offset;
         }
